+++ ios/Classes/AudioRecorder.swift
@@ -9,7 +9,7 @@ public class AudioRecorder: NSObject, AVAudioRecorderDelegate{
     
     public func startRecording(_ result: @escaping FlutterResult,_ path: String?,_ encoder : Int?,_ sampleRate : Int?,_ bitRate : Int?,_ fileNameFormat: String){
         let settings = [
-            AVEncoderBitRateKey: bitRate ?? 65536,
+            AVEncoderBitRateKey: bitRate ?? 32768,
             AVFormatIDKey: getEncoder(encoder ?? 0),
             AVSampleRateKey: sampleRate ?? 16000,
             AVNumberOfChannelsKey: 1,