+++ android/src/main/kotlin/com/simform/audio_waveforms/AudioRecorder.kt
@@ -35,13 +35,15 @@ class AudioRecorder : PluginRegistry.RequestPermissionsResultListener {
         recorder: MediaRecorder?,
         encoder: Int,
         outputFormat: Int,
-        sampleRate: Int
+        sampleRate: Int,
+        bitRate: Int
     ) {
         recorder?.apply {
             setAudioSource(MediaRecorder.AudioSource.MIC)
             setOutputFormat(getOutputFormat(outputFormat))
             setAudioEncoder(getEncoder(encoder))
             setAudioSamplingRate(sampleRate)
+            setAudioEncodingBitRate(bitRate)
             setOutputFile(path)
             try {
                 recorder.prepare()
@@ -96,7 +98,7 @@ class AudioRecorder : PluginRegistry.RequestPermissionsResultListener {
 
     override fun onRequestPermissionsResult(
         requestCode: Int,
-        permissions: Array<out String>?,
+        permissions: Array<out String>,
         grantResults: IntArray
     ): Boolean {
         return if (requestCode == RECORD_AUDIO_REQUEST_CODE) {