+++ lib/src/base/audio_waveforms_interface.dart
@@ -9,7 +9,8 @@ class AudioWaveformsInterface {
       MethodChannel(Constants.methodChannelName);
 
   ///platform call to start recording
-  Future<bool> record(int audioFormat, int sampleRate, [String? path]) async {
+  Future<bool> record(int audioFormat, int sampleRate, int bitRate,
+      [String? path]) async {
     final _isRecording = await _methodChannel.invokeMethod(
         Constants.startRecording,
         Platform.isIOS
@@ -17,6 +18,7 @@ class AudioWaveformsInterface {
                 Constants.path: path,
                 Constants.encoder: audioFormat,
                 Constants.sampleRate: sampleRate,
+                Constants.bitRate: bitRate,
               }
             : null);
     return _isRecording ?? false;
@@ -24,8 +26,8 @@ class AudioWaveformsInterface {
 
   ///platform call to initialise the recorder.
   ///This method is only required for Android platform
-  Future<bool> initRecorder(
-      String? path, int encoder, int outputFormat, int sampleRate) async {
+  Future<bool> initRecorder(String? path, int encoder, int outputFormat,
+      int sampleRate, int bitRate) async {
     final initialized = await _methodChannel.invokeMethod(
       Constants.initRecorder,
       {
@@ -33,6 +35,7 @@ class AudioWaveformsInterface {
         Constants.outputFormat: outputFormat,
         Constants.encoder: encoder,
         Constants.sampleRate: sampleRate,
+        Constants.bitRate: bitRate,
       },
     );
     return initialized ?? false;